requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
spacy
nltk
pandas
vaderSentiment
streamlit
